refactor: Waiting Query Service 분리
